fix ints fantasy

+++ magnate/fantasy.py
@@ -294,7 +294,7 @@ def apply_fantasy_event(game: Game, user: CustomUser , fantasy_event: FantasyEve
 
         ratio_to_add = fantasy_event.value
         previous_money = game.money[str(user.pk)]
-        game.money[str(user.pk)] = game.money[str(user.pk)] * (1 + ratio_to_add/100)
+        game.money[str(user.pk)] = int(game.money[str(user.pk)] * (1 + ratio_to_add/100))
         game.save()
         stats = PlayerGameStatistic.objects.get(user=user,game=game)
         stats.won_money += game.money[str(user.pk)] - previous_money
@@ -328,7 +328,7 @@ def apply_fantasy_event(game: Game, user: CustomUser , fantasy_event: FantasyEve
 
         ratio_to_sub = fantasy_event.value
         previous_money = game.money[str(user.pk)]
-        end_money = game.money[str(user.pk)] * (1 - ratio_to_sub/100)
+        end_money = int(game.money[str(user.pk)] * (1 - ratio_to_sub/100))
         game.parking_money += game.money[str(user.pk)] - end_money
         game.money[str(user.pk)] = end_money        
         game.save()
